Request system/requirements.txt
streamlit>=1.28.0
pandas==1.5.3
python-docx>=0.8.11
openpyxl>=3.1.0
xlrd>=2.0.1
numpy<2

